third_party/WebKit/LayoutTests/screen_orientation/screenorientation-api.html

<!DOCTYPE html>
<html>
<body>
<script src="../resources/js-test.js"></script>
<script>
description("Validates the Screen Orientation API is exposed.");

shouldBeDefined("screen.lockOrientation");
shouldBeDefined("screen.unlockOrientation");
shouldBeDefined("screen.orientation");
shouldBeDefined("window.onorientationchange");
</script>
</body>
</html>
